# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: Approval queue (approvers/admins) or own requests (editors)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApprovalStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalList"
        "400":
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/settings:
    get:
      operationId: getApprovalSettings
      tags: [approvals]
      summary: Whether approvals are required in this workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalSettings"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateApprovalSettings
      tags: [approvals]
      summary: Turn required approvals on or off (owner/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalSettingsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalSettings"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/{approvalId}:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getApproval
      tags: [approvals]
      summary: One approval with its comments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDetail"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{postId}/approval:
    parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: requestApproval
      tags: [approvals]
      summary: Ask for approval of a draft or scheduled post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Post not approvable or already pending/approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/{approvalId}/approve:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: approveApproval
      tags: [approvals]
      summary: Approve a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Not in a state that allows this
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/{approvalId}/reject:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: rejectApproval
      tags: [approvals]
      summary: Reject a post (note required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Not in a state that allows this
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/{approvalId}/request-changes:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: requestApprovalChanges
      tags: [approvals]
      summary: Ask for changes (note required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Not in a state that allows this
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/{approvalId}/withdraw:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: withdrawApproval
      tags: [approvals]
      summary: Withdraw your request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalWithdrawInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Not in a state that allows this
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approvals/{approvalId}/comments:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: addApprovalComment
      tags: [approvals]
      summary: Comment on an approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalCommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalComment"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected, changes_requested, withdrawn]
    ApprovalPostSummary:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        status: { type: string }
        scheduledAt: { type: ["string", "null"], format: date-time }
      required: [id, content, status, scheduledAt]
    Approval:
      type: object
      properties:
        id: { type: string }
        postId: { type: string }
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        requestedBy: { type: ["string", "null"] }
        requestedByName: { type: ["string", "null"] }
        requestedAt: { type: string, format: date-time }
        decidedBy: { type: ["string", "null"] }
        decidedByName: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"], format: date-time }
        note: { type: ["string", "null"] }
        post:
          oneOf:
            - $ref: "#/components/schemas/ApprovalPostSummary"
            - type: "null"
      required: [id, postId, status, requestedBy, requestedByName, requestedAt, decidedBy, decidedByName, decidedAt, note, post]
    ApprovalList:
      type: object
      properties:
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        scope: { type: string, enum: [queue, own] }
      required: [approvals, scope]
    ApprovalComment:
      type: object
      properties:
        id: { type: string }
        userId: { type: ["string", "null"] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, userId, body, createdAt]
    ApprovalDetailComment:
      type: object
      properties:
        id: { type: string }
        userId: { type: ["string", "null"] }
        author: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, userId, author, body, createdAt]
    ApprovalDetail:
      allOf:
        - $ref: "#/components/schemas/Approval"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/ApprovalDetailComment"
          required: [comments]
    ApprovalSettings:
      type: object
      properties:
        required: { type: boolean }
        updatedBy: { type: ["string", "null"] }
        updatedAt: { type: ["string", "null"], format: date-time }
      required: [required, updatedBy, updatedAt]
    ApprovalSettingsInput:
      type: object
      properties:
        required: { type: boolean }
      required: [required]
    ApprovalRequestInput:
      type: object
      properties:
        note: { type: ["string", "null"] }
    ApprovalDecisionInput:
      type: object
      properties:
        note: { type: ["string", "null"] }
    ApprovalWithdrawInput:
      type: object
      properties:
        note: { type: ["string", "null"] }
    ApprovalCommentInput:
      type: object
      properties:
        body: { type: string }
      required: [body]
...